+++ src/main/java/com/metsci/laproc/tools/GraphOptionsPanel.java
@@ -1,12 +1,10 @@
-package com.metsci.laproc.display;
+package com.metsci.laproc.tools;
 
 import com.metsci.glimpse.docking.View;
-import com.metsci.laproc.ActionHandlers.*;
+import com.metsci.laproc.action.*;
 import com.metsci.laproc.plotting.Graph;
 import com.metsci.laproc.utils.IAction;
-import com.metsci.laproc.application.DataReference;
 import com.metsci.laproc.pointmetrics.ParametricFunction;
-import com.metsci.laproc.utils.IObserver;
 
 import javax.swing.*;
 import java.awt.*;
@@ -18,7 +16,7 @@ import java.util.*;
  * Panel for selecting graph options
  * Created by porterjc on 10/26/2016.
  */
-public class GraphOptionsPanel implements ITool, IObserver<DataReference>{
+public class GraphOptionsPanel implements ITool, DataObserver{
     private JPanel panel;
     private JComboBox xaxis;
     private JComboBox yaxis;
